Move things where they should be
